refactor(testing): decompose e2e mock_fixtures.py into a package
Split tests/e2e/mock_fixtures.py (928 lines) into tests/e2e/mock_fixtures/
by concern (manifests, scheduler, telemetry, sessions/config, shared
constants), matching the existing test_utils package+re-export pattern.
__init__.py re-exports the full original public surface so conftest.py
and test_app_detail.py need no import changes. No behavior change.

Closes #1577
